reports/requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.2
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
Deprecated==1.2.14
et-xmlfile==1.1.0
firebase-admin==6.5.0
fonttools==4.54.1
frozenlist==1.4.1
future==1.0.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
kiwisolver==1.4.7
line-bot-sdk==3.13.0
matplotlib==3.9.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.13.1
